requirements.txt
redshift-connector
flask
flask-cors
strands-agents
requests
configparser